net: add wait_for_nic function
This commit adds a function called wait_for_nic(char *nic_name).
This function will check if a networ card exist in the the virtual
machine, if not it will wait until the device is detected.

Use udev events will prevent race conditions trying to use
a network card that is not detected yet.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>