net: Find network interface name from MAC address
In case of virtual machines, it is complicated to know what will
be the network interface name created by the hypervisor. But in
case we give a MAC address related to the TAP interface passed to
the hypervisor, we will be able to find this same address from
the agent running inside the VM.

That's why this patch provides a way to find a network interface
name from a MAC address, avoiding to try to predict what will be
the network interface name created.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>