Move containers to their own PID namespace
This commit moves the creation of the PID namespace from the pod level
to the container level thus aligning more closely with the k8s/docker
expectation of what is shared.

By moving the container to its own namespace this means the hyperstart
init is not PID 1 inside the containers thus preventing escape from the
container fs to the pod fs via /proc/1/cwd.

This change works by moving the switch to the PID namespace and the
subsequent fork from hyper_enter_sandbox() to hyper_do_exec_cmd(). This
is necessary as hyper_enter_sandbox() is called with a temporary process
to setup the container rootfs and the first process created in a
namespace must persist.

The first time the container is entered a new PID namespace is created
otherwise the existing one is entered. The namespace fd is acquired by
hyper_run_process() by looking up the namespace on the pid of the first
process launched in the container.

When creating a new PID namespace (not for joining) it is necessary to
remount /proc so that only the new processes are visible.

Fixes: 01org/cc-oci-runtime#768

Signed-off-by: Rob Bradford <robert.bradford@intel.com>